separated di, domain, data into individual module to achieve the modularization.
