+++ app/build.gradle.kts
@@ -43,6 +43,9 @@ android {
 
 dependencies {
 
+    implementation(project(":di"))
+    implementation(project(":domain"))
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
