--- app/src/main/java/com/lbg/assessment/data/ApiConstants.kt
@@ -1,6 +1,0 @@
-package com.lbg.assessment.data
-
-object ApiConstants {
-    const val BASE_URL = "https://jsonplaceholder.typicode.com/"
-    const val GET_ALL_USERS_EP = "users"
-}
