--- app/src/main/java/com/lbg/assessment/di/AppModule.kt
@@ -1,36 +1,0 @@
-package com.lbg.assessment.di
-
-import com.lbg.assessment.data.ApiConstants.BASE_URL
-import com.lbg.assessment.data.UserApiService
-import com.lbg.assessment.data.UserRepositoryImpl
-import com.lbg.assessment.domain.repository.UserRepository
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object AppModule {
-
-    @Provides
-    @Singleton
-    fun providesRetrofit(): Retrofit =
-        Retrofit.Builder().baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-
-    @Provides
-    @Singleton
-    fun providesApiService(retrofit: Retrofit): UserApiService =
-        retrofit.create(UserApiService::class.java)
-
-    @Provides
-    @Singleton
-    fun providesUserRepository(userApiService: UserApiService): UserRepository =
-        UserRepositoryImpl(userApiService)
-
-}
