--- app/src/main/java/com/lbg/assessment/domain/model/User.kt
@@ -1,8 +1,0 @@
-package com.lbg.assessment.domain.model
-
-data class User(
-    val id: Int,
-    val name: String,
-    val email: String,
-    val phone: String
-)
