--- app/src/main/java/com/lbg/assessment/domain/repository/UserRepository.kt
@@ -1,7 +1,0 @@
-package com.lbg.assessment.domain.repository
-
-import com.lbg.assessment.domain.model.User
-
-interface UserRepository {
-    suspend fun fetchUsers(): List<User>
-}
