+++ data/src/main/java/com/lbg/data/ApiConstants.kt
@@ -1,0 +1,6 @@
+package com.lbg.data
+
+object ApiConstants {
+    const val BASE_URL = "https://jsonplaceholder.typicode.com/"
+    const val GET_ALL_USERS_EP = "users"
+}
