+++ data/src/main/java/com/lbg/data/service/UserApiService.kt
@@ -1,0 +1,10 @@
+package com.lbg.data.service
+
+import com.lbg.data.ApiConstants
+import retrofit2.http.GET
+import com.lbg.domain.model.User
+
+interface UserApiService {
+    @GET(ApiConstants.GET_ALL_USERS_EP)
+    suspend fun getUsers(): List<User>
+}
