+++ di/src/main/java/com/lbg/di/AppModule.kt
@@ -1,0 +1,36 @@
+package com.lbg.di
+
+import com.lbg.data.ApiConstants.BASE_URL
+import com.lbg.data.repository.UserRepositoryImpl
+import com.lbg.data.service.UserApiService
+import com.lbg.domain.repository.UserRepository
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+
+    @Provides
+    @Singleton
+    fun providesRetrofit(): Retrofit =
+        Retrofit.Builder().baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+
+    @Provides
+    @Singleton
+    fun providesApiService(retrofit: Retrofit): UserApiService =
+        retrofit.create(UserApiService::class.java)
+
+    @Provides
+    @Singleton
+    fun providesUserRepository(userApiService: UserApiService): UserRepository =
+        UserRepositoryImpl(userApiService)
+
+}
