+++ domain/src/main/java/com/lbg/domain/model/User.kt
@@ -1,0 +1,8 @@
+package com.lbg.domain.model
+
+data class User(
+    val id: Int,
+    val name: String,
+    val email: String,
+    val phone: String
+)
