+++ domain/src/main/java/com/lbg/domain/repository/UserRepository.kt
@@ -1,0 +1,7 @@
+package com.lbg.domain.repository
+
+import com.lbg.domain.model.User
+
+interface UserRepository {
+    suspend fun fetchUsers(): List<User>
+}
