+++ gradle/libs.versions.toml
@@ -14,6 +14,8 @@ hiltNavigationCompose = "1.2.0"
 junitJunit = "4.12"
 coroutines = "1.7.0"
 mockito = "4.5.1"
+appcompat = "1.7.0"
+material = "1.12.0"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -38,6 +40,8 @@ androidx-hilt-navigation-compose = { group = "androidx.hilt", name = "hilt-navig
 junit-junit = { group = "junit", name = "junit", version.ref = "junitJunit" }
 org-mockito = { group = "org.mockito", name = "mockito-core", version.ref = "mockito" }
 kotlinx-coroutines-test = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-test", version.ref = "coroutines" }
+androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
+material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
@@ -45,4 +49,6 @@ kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 kotlin-compose = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
 kotlin-kapt = { id = "org.jetbrains.kotlin.kapt", version.ref = "kotlin" }
 hilt-android = { id = "com.google.dagger.hilt.android", version.ref = "daggerHilt" }
+android-library = { id = "com.android.library" }
+kotlin-android-only = { id = "kotlin-android" }
