+++ settings.gradle.kts
@@ -21,4 +21,6 @@ dependencyResolutionManagement {
 
 rootProject.name = "Client-assessment-compose"
 include(":app")
- 
+include(":di")
+include(":domain")
+include(":data")
